Reduce surface area of global Tarjan implementation
This commit makes the interface to function tarjan() more slim in
preparation of creating a local version.  In particular, we
introduce an abstract datatype that captures SCCs and that can be
queried through dedicated operations.  We also make tarjan() itself
into a higher-level manager that calls into a computational routine.

Update callers accordingly.