Separate SCC capture out to helper function
This commit switches the procedure for extracting a single SCC from
the component stack to using indexed access keyed off scc->start.  In
particular, we use the equivalent of C++'s std::vector<>::push_back()
from the end pointer of a component and record the new vertices while
advancing this end pointer.

In the process move this task out to separate helper function,
capture_scc(), to reduce the number of moving parts in the local DFS
vertex loop.