Refactor innards of local search to separate helpers
This commit introduces three helper functions

    - discover_vertex()
    - dfs_from_descendant()
    - complete_dfs_from_vertex()

that each handle one aspect of the combined DFS and SCC handling.

Function discover_vertex() handles the case of pushing a previously
unknown vertex onto the component stack, initialising the discovery
times and computing the number of descendants of the vertex (stored
in 'status').

Function dfs_from_descendant() initiates a DFS from the next
descendant of a vertex while maintaining the low-link count of the
parent vertex.

Function complete_dfs_from_vertex() removes a particular vertex from
the vertex stack and records a strong component if that vertex's
low-link count is the same as its discovery time.