Tarjan_global(): Switch to sparse start point processing
This commit introduces a new helper function, tarjan_find_start(),
that performs a linear scan over the 'status' array from a
particular seed point (the current start point) looking for the next
undiscovered vertex.

This makes the global management loop in tarjan_global() simpler
because the conditional can be moved into the for() condition.