Tarjan: Add ability to reverse order of SCCs
This is useful because Tarjan's algorithm constructs a reverse
ordering of the SCCs in the directed graph (downstream neighbours).

In the application of computing a volume-averaged ToF for all cells,
this does not matter because we can use the inflow graph directly.
When doing local analysis however, we need to use the outflow graph
and then the result comes back in the reverse order of what we
expecet.  The alternative is then either to visit the components in
reverse (descending) order (tarjan_get_strongcomponent()) or to call

    tarjan_reverse_sccresult()

and visit the components in ascending order.

Test the component reversal code in all existing Tarjan tests.