Fix multi-cell solve and make max tof work like in MRST.
A bug was discovered with multi-cell solves: writing to the
upwind_contrib_ for downstream cells during a multi-cell solve
was wrong, and not easy to fix. The approach used was changed so
that upwind contribution was computed before computing tof in
solveSingleCell(), instead of written to after computing tod.
The new approach is also simpler, not requiring a member array
for the upwind contributions.

Treatment of maximum time-of-flight has changed to match MRSTs
somewhat convoluted approach: first cap only when local inflow
fill time is above the max, post-process to cap every cell after
the solution step proper (done in solve()).