Only add connection if flux strictly different from zero.
While not a bug as such, adding zero fluxes has the following
unfortunate consequences:
 - creates unnecessary connected components (triggering unnecessary multi-cell solves)
 - causes the domains of local solutions to include unnecessary cells and components

The second is the most problematic, although filtering on zero tracer values could also
deal with it. The current fix is preferred though, as it also improves performance and
makes the algorithms perform more intuitively (i.e. not visiting unnecessary cells).