Add Measure of Resilience to Local ToF Calculations
In particular, guard against all upstream neighbours of cell having
too little inflow to get a well-defined tracer value.  In the case
of a local solve--meaning we cover only those cells that are
downstream from a particular set of starting points--it is possible
that all of cell's upstream neighbours are either unreachable or get
treated to the "little inflow" early return from solveSingleCell().
In this case, both 'upwind_tof_contrib' and 'upwind_tracer_contrib'
become zero, and the subsequent ToF calculation leads to 0/0
arithmetic operation which generates a NaN value that propagates to
its downstream neighbours and contaminates the numerical solution.

Guard against this situation by explictly checking that the cell's
tracer value is positive before attempting the division.  We
arbitrarily assign maximum time-of-flight to cells that get zero
tracer concentrations in this fashion.