Change argument of Toolbox::assignInflowFlux().
We change from a (cell index -> value) association to a
nested (ID -> (cell index -> value)) association. This is
in preparation for handling cases when multiple wells can
connect to the same cell.

Also the interface of injectorProducerPairFlux() has changed
to expect the same as assignInflowFlux() this incidentally
allowed further simplifying the interface and implementation of that
function.

Finally, tests that invoke the toolbox have been modified to use the
new API. Note that the failing test introduced in the flow diagnostics
is still failing (the feature is still not implemented).